Fixed CA name and description, moved HoleESP and HoleSnap to Combat category
